+++ src/ItemManager.ts
@@ -1,39 +1,6 @@
 import { MULTI, GROUPKEY_ATT, IGNORE_CLASSNAME, DUMMY_POSITION } from "./consts";
 import { $, toArray, isUndefined, assign } from "./utils";
-import { CursorType, IInfiniteGridItemElement, ISize, IPosition, IJQuery } from "./types";
-
-/**
- * Item Interface
- * @ko Item Interface
- * @memberof eg.InfiniteGrid
- * @typedef
- * @property - Key in group containing item <ko>item을 포함하고 있는 group의 key</ko>
- * @property - html of element in item <ko>item의 element의 html</ko>
- * @property {HTMLElement} - Element for the item <ko>Element for the item</ko>
- * @property - Size at first rendering <ko>처음 렌더링 했을 때의 사이즈</ko>
- * @property - Currently seen the size of element <ko>현재 보여지는 Element의 사이즈</ko>
- * @property - Position and size shown in layout <ko>레이아웃할 때 보여지는 포지션과 사이즈</ko>
- * @property - Position and size that was shown when I was layouted <ko>전에 레이아웃했을 때 보였었던 포지션과 사이즈</ko>
- */
-export interface IInfiniteGridItem {
-	groupKey: string | number;
-	content: string;
-	el?: IInfiniteGridItemElement;
-	orgSize?: ISize;
-	size?: ISize;
-	rect?: Partial<ISize & IPosition>;
-	prevRect?: Partial<ISize & IPosition>;
-	[key: string]: any;
-}
-export interface IInfiniteGridGroup {
-	groupKey: string | number;
-	items: IInfiniteGridItem[];
-	outlines: { start: number[], end: number[] };
-}
-
-export interface IItemManagerStatus {
-	_data: IInfiniteGridGroup[];
-}
+import { CursorType, IJQuery, IInfiniteGridGroup, IInfiniteGridItem, IItemManagerStatus } from "./types";
 
 export default class ItemManager {
 	public static from(
