+++ src/types.ts
@@ -1,5 +1,3 @@
-import { IInfiniteGridItem, IInfiniteGridGroup } from "./ItemManager";
-
 export type CursorType = "start" | "end";
 export type SizeType = "width" | "height";
 export type PositionType = "left" | "top";
@@ -9,6 +7,97 @@ export type InnerSizeType = "innerWidth" | "innerHeight";
 export type ClientSizeType = "clientWidth" | "clientHeight";
 export type OffsetSizeType = "offsetWidth" | "offsetHeight";
 export type ScrollSizeType = "scrollWidth" | "scrollHeight";
+
+export interface IInfiniteGridGroup {
+	groupKey: string | number;
+	items: IInfiniteGridItem[];
+	outlines: { start: number[], end: number[] };
+}
+
+export interface IItemManagerStatus {
+	_data: IInfiniteGridGroup[];
+}
+
+export interface IInfiniteStatus {
+	startCursor: number;
+	endCursor: number;
+	size: number;
+}
+
+export interface IDOMRendererSize {
+	container?: number;
+	view?: number;
+	viewport?: number;
+	item?: ISize;
+}
+
+export interface IDOMRendererStatus {
+	cssText: string;
+	_size: IDOMRendererSize;
+}
+
+export interface IWatchStatus {
+	_prevPos: number;
+	scrollPos: number;
+}
+
+/**
+ * The object of data to be sent to an event
+ * @ko 이벤트에 전달되는 데이터 객체
+ * @memberof eg.InfiniteGrid
+ * @typedef
+ * @property - Appending card's image element.<ko>추가 되는 카드의 이미지 엘리먼트</ko>
+ * @property - The item's element with error images.<ko>에러난 이미지를 가지고 있는 아이템의 엘리먼트</ko>
+ * @property - The items being added.<ko>화면에 추가중인 아이템들</ko>
+ * @property - The item with error images.<ko>에러난 이미지를 가지고 있는 아이템</ko>
+ * @property - The item's index with error images.<ko>에러난 이미지를 가지고 있는 아이템의 인덱스</ko>
+ * @property - In the imageError event, this method expects to remove the error image.<ko>이미지 에러 이벤트에서 이 메서드는 에러난 이미지를 삭제한다.</ko>
+ * @property - In the imageError event, this method expects to remove the item with the error image.<ko>이미지 에러 이벤트에서 이 메서드는 에러난 이미지를 가지고 있는 아이템을 삭제한다.</ko>
+ * @property - In the imageError event, this method expects to replace the error image's source or element.<ko>이미지 에러 이벤트에서 이 메서드는 에러난 이미지의 주소 또는 엘리먼트를 교체한다.</ko>
+ * @property - In the imageError event, this method expects to replace the item's contents with the error image.<ko>이미지 에러 이벤트에서 이 메서드는 에러난 이미지를 가지고 있는 아이템의 내용을 교체한다.</ko>
+ * @example
+	ig.on("imageError", e => {
+	  e.remove();
+	  e.removeItem();
+	  e.replace("http://...jpg");
+	  e.replace(imageElement);
+	  e.replaceItem("item html");
+	});
+ */
+export interface IErrorCallbackOptions {
+	target: HTMLImageElement;
+	element: HTMLElement;
+	items: IInfiniteGridItem[];
+	item: IInfiniteGridItem;
+	itemIndex: number;
+	replace: (src: string) => void;
+	replaceItem: (content: string) => void;
+	remove: () => void;
+	removeItem: () => void;
+}
+/**
+ * Item Interface
+ * @ko Item Interface
+ * @memberof eg.InfiniteGrid
+ * @typedef
+ * @property - Key in group containing item <ko>item을 포함하고 있는 group의 key</ko>
+ * @property - html of element in item <ko>item의 element의 html</ko>
+ * @property {HTMLElement} - Element for the item <ko>Element for the item</ko>
+ * @property - Size at first rendering <ko>처음 렌더링 했을 때의 사이즈</ko>
+ * @property - Currently seen the size of element <ko>현재 보여지는 Element의 사이즈</ko>
+ * @property - Position and size shown in layout <ko>레이아웃할 때 보여지는 포지션과 사이즈</ko>
+ * @property - Position and size that was shown when I was layouted <ko>전에 레이아웃했을 때 보였었던 포지션과 사이즈</ko>
+ */
+export interface IInfiniteGridItem {
+	groupKey: string | number;
+	content: string;
+	el?: IInfiniteGridItemElement;
+	orgSize?: ISize;
+	size?: ISize;
+	rect?: Partial<ISize & IPosition>;
+	prevRect?: Partial<ISize & IPosition>;
+	[key: string]: any;
+}
 
 // see https://github.com/Microsoft/TypeScript/issues/27024#issuecomment-421529650
 export type Equals<X, Y, A, B = never> =
@@ -93,3 +182,7 @@ export interface ILayout {
 	setSize(size: number): this;
 	layout(groups: IInfiniteGridGroup[], outline: number[]): this;
 }
+
+export type WindowMockType = {
+	[P in keyof Window]?: Window[P] extends (...args: any[]) => any ? Window[P] : Partial<Window[P]>
+};
